Move ContentType from content_type package to core package
