+++ shared/src/commonMain/kotlin/org/hyperskill/app/comments/domain/model/Comment.kt
@@ -3,7 +3,7 @@ package org.hyperskill.app.comments.domain.model
 import kotlinx.datetime.Instant
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import org.hyperskill.app.content_type.domain.model.ContentType
+import org.hyperskill.app.core.domain.model.ContentType
 
 @Serializable
 data class Comment(
