+++ shared/src/commonMain/kotlin/org/hyperskill/app/learning_activities/presentation/mapper/LearningActivityTargetViewActionMapper.kt
@@ -1,6 +1,6 @@
 package org.hyperskill.app.learning_activities.presentation.mapper
 
-import org.hyperskill.app.content_type.domain.model.ContentType
+import org.hyperskill.app.core.domain.model.ContentType
 import org.hyperskill.app.learning_activities.domain.model.LearningActivity
 import org.hyperskill.app.learning_activities.domain.model.LearningActivityType
 import org.hyperskill.app.learning_activities.presentation.model.LearningActivityTargetViewAction
