+++ shared/src/commonMain/kotlin/org/hyperskill/app/likes/data/repository/LikesRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package org.hyperskill.app.likes.data.repository
 
-import org.hyperskill.app.content_type.domain.model.ContentType
+import org.hyperskill.app.core.domain.model.ContentType
 import org.hyperskill.app.likes.data.source.LikesRemoteDataSource
 import org.hyperskill.app.likes.domain.model.Like
 import org.hyperskill.app.likes.domain.model.LikeSubject
