+++ shared/src/commonMain/kotlin/org/hyperskill/app/likes/domain/model/Like.kt
@@ -2,7 +2,7 @@ package org.hyperskill.app.likes.domain.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import org.hyperskill.app.content_type.domain.model.ContentType
+import org.hyperskill.app.core.domain.model.ContentType
 
 @Serializable
 data class Like(
