+++ shared/src/commonMain/kotlin/org/hyperskill/app/search_results/domain/model/SearchResult.kt
@@ -2,7 +2,7 @@ package org.hyperskill.app.search_results.domain.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import org.hyperskill.app.content_type.domain.model.ContentType
+import org.hyperskill.app.core.domain.model.ContentType
 
 @Serializable
 data class SearchResult(
