+++ shared/src/commonMain/kotlin/org/hyperskill/app/study_plan/domain/analytic/StudyPlanClickedActivityHyperskillAnalyticEvent.kt
@@ -5,7 +5,7 @@ import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticEve
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticPart
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticRoute
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticTarget
-import org.hyperskill.app.content_type.domain.model.ContentType
+import org.hyperskill.app.core.domain.model.ContentType
 import ru.nobird.app.core.model.mapOfNotNull
 
 /**
